Convert standalone controllers in Exploration Editor page to component directives (#6843)
* Exploration editor tab

* Preview Tab

* Statistics Tab

* Settings Tab

* History Tab

* Fix karma of History Tab and Exploration Editor tab

* Feedback Tab

* Revert "History Tab"

This reverts commit 495007f627b2e2df38bee7211d3f1df67c242da3.

* Revert History tab spec changes

* Fix tests and restore HIstoryTabDirective

* Fix tests and lint

* Fix e2e test publicationAndLibrary

* Fix lint

* Fix lint

* Break line

* Maintained consistency
