+++ core/templates/dev/head/pages/exploration_editor/ExplorationEditor.ts
@@ -106,21 +106,21 @@ require(
   'pages/exploration_editor/ExplorationSaveAndPublishButtonsDirective.ts');
 require('pages/exploration_editor/ExplorationTitleEditorDirective.ts');
 require('pages/exploration_editor/ParamChangesEditorDirective.ts');
-require('pages/exploration_editor/editor_tab/ExplorationEditorTab.ts');
+require('pages/exploration_editor/editor_tab/ExplorationEditorTabDirective.ts');
 require('pages/exploration_editor/editor_tab/ExplorationGraphDirective.ts');
 require(
   'pages/exploration_editor/editor_tab/UnresolvedAnswersOverviewDirective.ts');
-require('pages/exploration_editor/feedback_tab/FeedbackTab.ts');
+require('pages/exploration_editor/feedback_tab/FeedbackTabDirective.ts');
 require('pages/exploration_editor/feedback_tab/ThreadTableDirective.ts');
-require('pages/exploration_editor/history_tab/HistoryTab.ts');
+require('pages/exploration_editor/history_tab/HistoryTabDirective.ts');
 require(
   'pages/exploration_editor/improvements_tab/ImprovementsTabDirective.ts');
-require('pages/exploration_editor/preview_tab/PreviewTab.ts');
-require('pages/exploration_editor/settings_tab/SettingsTab.ts');
+require('pages/exploration_editor/preview_tab/PreviewTabDirective.ts');
+require('pages/exploration_editor/settings_tab/SettingsTabDirective.ts');
 require('pages/exploration_editor/statistics_tab/PieChartDirective.ts');
 require(
   'pages/exploration_editor/statistics_tab/PlaythroughIssuesDirective.ts');
-require('pages/exploration_editor/statistics_tab/StatisticsTab.ts');
+require('pages/exploration_editor/statistics_tab/StatisticsTabDirective.ts');
 require('pages/exploration_editor/translation_tab/TranslationTabDirective.ts');
 require('pages/exploration_player/ConversationSkinDirective.ts');
 require('pages/exploration_player/ExplorationFooterDirective.ts');
