--- core/templates/dev/head/pages/exploration_editor/editor_tab/exploration_editor_tab.html
@@ -1,25 +1,0 @@
-<div ng-controller="ExplorationEditorTab" class="oppia-editor-cards-container">
-  <div class="oppia-editor-header">
-    <state-name-editor></state-name-editor>
-  </div>
-
-  <div ng-if="areParametersEnabled()">
-    <state-param-changes-editor></state-param-changes-editor>
-  </div>
-
-  <state-editor on-save-state-content="saveStateContent"
-                on-save-interaction-customization-args="saveInteractionCustomizationArgs"
-                on-save-interaction-id="saveInteractionId"
-                on-save-interaction-default-outcome="saveInteractionDefaultOutcome"
-                on-save-interaction-answer-groups="saveInteractionAnswerGroups"
-                on-save-solution="saveSolution"
-                on-save-hints="saveHints"
-                recompute-graph="recomputeGraph"
-                show-mark-all-audio-as-needing-update-modal-if-required="showMarkAllAudioAsNeedingUpdateModalIfRequired"
-                state-content-placeholder="getStateContentPlaceholder()"
-                navigate-to-state="navigateToState"
-                add-state="addState"
-                refresh-warnings="refreshWarnings"
-                interaction-is-shown="interactionIsShown">
-  </state-editor>
-</div>
