+++ core/templates/dev/head/pages/exploration_editor/exploration_editor.html
@@ -68,32 +68,26 @@
         {% include 'pages/exploration_editor/editor_tab/editor_tab.html' %}
       </div>
 
-      <div ng-if="getActiveTabName() === 'translation'">
-        <translation-tab></translation-tab>
-      </div>
+      <translation-tab ng-if="getActiveTabName() === 'translation'">
+      </translation-tab>
       <!-- This is an ng-if, so that the preview loads the latest version of the exploration each time the tab is accessed. -->
-      <div ng-if="getActiveTabName() === 'preview'">
-        {% include 'pages/exploration_editor/preview_tab/preview_tab.html' %}
-      </div>
+      <preview-tab ng-if="getActiveTabName() === 'preview'">
+      </preview-tab>
 
-      <div ng-show="getActiveTabName() === 'stats'">
-        {% include 'pages/exploration_editor/statistics_tab/statistics_tab.html' %}
-      </div>
+      <statistics-tab ng-show="getActiveTabName() === 'stats'">
+      </statistics-tab>
 
       <improvements-tab ng-if="getActiveTabName() === 'improvements'">
       </improvements-tab>
 
-      <div ng-show="getActiveTabName() === 'settings'">
-        {% include 'pages/exploration_editor/settings_tab/settings_tab.html' %}
-      </div>
+      <settings-tab ng-show="getActiveTabName() === 'settings'" current-user-is-admin="currentUserIsAdmin" current-user-is-moderator="currentUserIsModerator">
+      </settings-tab>
 
-      <div ng-show="getActiveTabName() === 'history'">
-        {% include 'pages/exploration_editor/history_tab/history_tab.html' %}
-      </div>
+      <history-tab ng-show="getActiveTabName() === 'history'">
+      </history-tab>
 
-      <div ng-show="getActiveTabName() === 'feedback'">
-        {% include 'pages/exploration_editor/feedback_tab/feedback_tab.html' %}
-      </div>
+      <feedback-tab ng-show="getActiveTabName() === 'feedback'">
+      </feedback-tab>
     </div>
   </div>
