+++ core/templates/dev/head/pages/exploration_editor/feedback_tab/FeedbackTabDirective.ts
@@ -1,0 +1,218 @@
+// Copyright 2014 The Oppia Authors. All Rights Reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS-IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+/**
+ * @fileoverview Controller for the exploration editor feedback tab.
+ */
+
+require('domain/state/StateObjectFactory.ts');
+require('domain/utilities/UrlInterpolationService.ts');
+require('pages/exploration_editor/ChangeListService.ts');
+require('pages/exploration_editor/ExplorationDataService.ts');
+require('pages/exploration_editor/ExplorationStatesService.ts');
+require('pages/exploration_editor/feedback_tab/ThreadDataService.ts');
+require('pages/exploration_editor/feedback_tab/ThreadStatusDisplayService.ts');
+require('pages/suggestion_editor/ShowSuggestionModalForEditorViewService.ts');
+require('services/AlertsService.ts');
+require('services/DateTimeFormatService.ts');
+require('services/UserService.ts');
+
+oppia.directive('feedbackTab', ['UrlInterpolationService', function(
+    UrlInterpolationService) {
+  return {
+    restrict: 'E',
+    scope: {},
+    bindToController: {},
+    templateUrl: UrlInterpolationService.getDirectiveTemplateUrl(
+      '/pages/exploration_editor/feedback_tab/feedback_tab_directive.html'),
+    controllerAs: '$ctrl',
+    controller: [
+      '$log', '$q', '$rootScope', '$uibModal',
+      'AlertsService', 'ChangeListService', 'DateTimeFormatService',
+      'ExplorationDataService', 'ExplorationStatesService',
+      'ShowSuggestionModalForEditorViewService',
+      'StateObjectFactory',
+      'ThreadDataService', 'ThreadStatusDisplayService',
+      'UrlInterpolationService', 'UserService',
+      'ACTION_ACCEPT_SUGGESTION', 'ACTION_REJECT_SUGGESTION',
+      function(
+          $log, $q, $rootScope, $uibModal,
+          AlertsService, ChangeListService, DateTimeFormatService,
+          ExplorationDataService, ExplorationStatesService,
+          ShowSuggestionModalForEditorViewService,
+          StateObjectFactory,
+          ThreadDataService, ThreadStatusDisplayService,
+          UrlInterpolationService, UserService,
+          ACTION_ACCEPT_SUGGESTION, ACTION_REJECT_SUGGESTION) {
+        var ctrl = this;
+        ctrl.STATUS_CHOICES = ThreadStatusDisplayService.STATUS_CHOICES;
+        ctrl.threadData = ThreadDataService.data;
+        ctrl.getLabelClass = ThreadStatusDisplayService.getLabelClass;
+        ctrl.getHumanReadableStatus = (
+          ThreadStatusDisplayService.getHumanReadableStatus);
+        ctrl.getLocaleAbbreviatedDatetimeString = (
+          DateTimeFormatService.getLocaleAbbreviatedDatetimeString);
+
+        ctrl.activeThread = null;
+        ctrl.userIsLoggedIn = null;
+        ctrl.threadIsUpdated = false;
+        $rootScope.loadingMessage = 'Loading';
+        var userInfoPromise = UserService.getUserInfoAsync();
+        userInfoPromise.then(function(userInfo) {
+          ctrl.userIsLoggedIn = userInfo.isLoggedIn();
+        });
+
+        // Initial load of the thread list on page load.
+        ctrl.tmpMessage = {
+          status: null,
+          text: ''
+        };
+        var _resetTmpMessageFields = function() {
+          ctrl.tmpMessage.status = ctrl.activeThread ?
+            ctrl.activeThread.status : null;
+          ctrl.tmpMessage.text = '';
+        };
+        ctrl.clearActiveThread = function() {
+          ctrl.activeThread = null;
+          _resetTmpMessageFields();
+        };
+        ctrl.clearActiveThread();
+        ThreadDataService.fetchFeedbackStats();
+        var threadPromise = ThreadDataService.fetchThreads();
+        $q.all([userInfoPromise, threadPromise]).then(function() {
+          $rootScope.loadingMessage = '';
+        });
+        // Fetches the threads again if any thread is updated.
+        ctrl.fetchUpdatedThreads = function() {
+          ThreadDataService.fetchThreads();
+          ctrl.threadData = ThreadDataService.data;
+          ctrl.threadIsUpdated = false;
+        };
+        ctrl.onBackButtonClicked = function() {
+          ctrl.clearActiveThread();
+          if (ctrl.threadIsUpdated) {
+            ctrl.fetchUpdatedThreads();
+          }
+        };
+
+        ctrl.showCreateThreadModal = function() {
+          $uibModal.open({
+            templateUrl: UrlInterpolationService.getDirectiveTemplateUrl(
+              '/pages/exploration_editor/feedback_tab/' +
+              'editor_create_feedback_thread_modal_directive.html'),
+            backdrop: true,
+            resolve: {},
+            controller: ['$scope', '$uibModalInstance', function(
+                $scope, $uibModalInstance) {
+              $scope.newThreadSubject = '';
+              $scope.newThreadText = '';
+
+              $scope.create = function(newThreadSubject, newThreadText) {
+                if (!newThreadSubject) {
+                  AlertsService.addWarning('Please specify a thread subject.');
+                  return;
+                }
+                if (!newThreadText) {
+                  AlertsService.addWarning('Please specify a message.');
+                  return;
+                }
+
+                $uibModalInstance.close({
+                  newThreadSubject: newThreadSubject,
+                  newThreadText: newThreadText
+                });
+              };
+
+              $scope.cancel = function() {
+                $uibModalInstance.dismiss('cancel');
+              };
+            }]
+          }).result.then(function(result) {
+            ThreadDataService.createNewThread(
+              result.newThreadSubject, result.newThreadText, function() {
+                ctrl.clearActiveThread();
+                AlertsService.addSuccessMessage('Feedback thread created.');
+              });
+          });
+        };
+
+        var _isSuggestionHandled = function() {
+          return ctrl.activeThread.isSuggestionHandled();
+        };
+
+        var _isSuggestionValid = function() {
+          return ExplorationStatesService.hasState(
+            ctrl.activeThread.getSuggestionStateName());
+        };
+
+        var _hasUnsavedChanges = function() {
+          return (ChangeListService.getChangeList().length > 0);
+        };
+
+        ctrl.getSuggestionButtonType = function() {
+          return (!_isSuggestionHandled() && _isSuggestionValid() &&
+                  !_hasUnsavedChanges() ? 'primary' : 'default');
+        };
+
+        // TODO(Allan): Implement ability to edit suggestions before applying.
+        ctrl.showSuggestionModal = function() {
+          ShowSuggestionModalForEditorViewService.showSuggestionModal(
+            ctrl.activeThread.suggestion.suggestionType,
+            {
+              activeThread: ctrl.activeThread,
+              setActiveThread: ctrl.setActiveThread,
+              isSuggestionHandled: _isSuggestionHandled,
+              hasUnsavedChanges: _hasUnsavedChanges,
+              isSuggestionValid: _isSuggestionValid
+            }
+          );
+        };
+
+        ctrl.addNewMessage = function(threadId, tmpText, tmpStatus) {
+          if (threadId === null) {
+            AlertsService.addWarning(
+              'Cannot add message to thread with ID: null.');
+            return;
+          }
+          if (!tmpStatus) {
+            AlertsService.addWarning('Invalid message status: ' + tmpStatus);
+            return;
+          }
+          ctrl.threadIsUpdated = true;
+          ctrl.messageSendingInProgress = true;
+          ThreadDataService.addNewMessage(
+            threadId, tmpText, tmpStatus, function() {
+              _resetTmpMessageFields();
+              ctrl.messageSendingInProgress = false;
+            }, function() {
+              ctrl.messageSendingInProgress = false;
+            });
+        };
+
+        ctrl.setActiveThread = function(threadId) {
+          ThreadDataService.fetchMessages(threadId);
+          ThreadDataService.markThreadAsSeen(threadId);
+          var allThreads = [].concat(
+            ctrl.threadData.feedbackThreads, ctrl.threadData.suggestionThreads);
+          for (var i = 0; i < allThreads.length; i++) {
+            if (allThreads[i].threadId === threadId) {
+              ctrl.activeThread = allThreads[i];
+              break;
+            }
+          }
+          ctrl.tmpMessage.status = ctrl.activeThread.status;
+        };
+      }
+    ]};
+}]);
