+++ core/templates/dev/head/pages/exploration_editor/history_tab/history_tab_directive.html
@@ -1,0 +1,50 @@
+<div class="container-fluid" style="word-wrap: break-word;">
+  <md-card class="oppia-editor-card">
+    <h3>List of Changes</h3>
+    <div ng-repeat="versionNumber in $ctrl.versionNumbersToDisplay" class="row">
+      <div class="col-sm-4 col-md-4 col-lg-4">
+        <input type="checkbox" name="compareVer" ng-model="$ctrl.versionCheckboxArray[versionNumber].selected" ng-init="init()" ng-click="$ctrl.changeSelectedVersions($event, versionNumber)" ng-disabled="$ctrl.isCheckboxDisabled(versionNumber)" ng-hide="$ctrl.comparisonsAreDisabled" class="protractor-test-history-checkbox-selector">
+        [<[versionNumber]>]
+        <strong><profile-link-text username="$ctrl.explorationVersionMetadata[versionNumber].committerId"></profile-link-text></strong>
+      </div>
+      <div class="col-sm-4 col-md-4 col-lg-4">
+        <[$ctrl.explorationVersionMetadata[versionNumber].commitMessage]>
+      </div>
+      <div class="col-sm-4 col-md-4 col-lg-4">
+        <span>
+          <em><[$ctrl.explorationVersionMetadata[versionNumber].createdOnStr]></em>
+          <span ng-if="versionNumber !== $ctrl.currentVersion && $ctrl.EditabilityService.isEditable()">
+            | <a href="#" class="protractor-test-revert-version" ng-click="$ctrl.showRevertExplorationModal(versionNumber)">Revert</a>
+          </span>
+          <span>
+            | <a href="#" ng-click="$ctrl.downloadExplorationWithVersion(versionNumber)" title="Download exploration as a zip file.">Download</a>
+          </span>
+        </span>
+      </div>
+    </div>
+
+    <br>
+
+    <div ng-hide="!$ctrl.explorationVersionMetadata || $ctrl.compareVersionsButtonIsHidden">
+      <button class="btn protractor-test-show-history-graph" ng-class="{'btn-success': $ctrl.areCompareVersionsSelected(), 'btn-default': !$ctrl.areCompareVersionsSelected()}" ng-click="$ctrl.compareSelectedVersions()" ng-disabled="!$ctrl.areCompareVersionsSelected()">Compare selected revisions</button>
+      <div class="version-count-prompt">
+        <[$ctrl.versionCountPrompt]>
+      </div>
+    </div>
+
+    <div ng-show="$ctrl.versionCheckboxArray.length > $ctrl.VERSIONS_PER_PAGE" style="margin-top: 30px" uib-pagination boundary-links="true" class="pagination-sm" max-size="5" items-per-page="$ctrl.VERSIONS_PER_PAGE" total-items="$ctrl.versionCheckboxArray.length" ng-model="$ctrl.displayedCurrentPageNumber" ng-change="$ctrl.computeVersionsToDisplay()"></div>
+  </md-card>
+
+  <md-card class="oppia-editor-card" style="position: relative;"
+           ng-if="$ctrl.diffData && !$ctrl.hideHistoryGraph && $ctrl.explorationVersionMetadata">
+    <h3>
+      Changes from version <[$ctrl.compareVersionMetadata.earlierVersion.versionNumber]>
+      to version <[$ctrl.compareVersionMetadata.laterVersion.versionNumber]>
+    </h3>
+
+    <version-diff-visualization diff-data="$ctrl.diffData"
+                                earlier-version-header="$ctrl.earlierVersionHeader"
+                                later-version-header="$ctrl.laterVersionHeader">
+    </version-diff-visualization>
+  </md-card>
+</div>
