--- core/templates/dev/head/pages/exploration_editor/statistics_tab/statistics_tab.html
@@ -1,38 +1,0 @@
-<div ng-controller="StatisticsTab" ng-cloak>
-  <md-card class="oppia-editor-card" ng-if="!loadingMessage">
-    <div ng-if="explorationHasBeenVisited">
-      <div class="oppia-statistics-display">
-        <h3>Exploration Completion Rate</h3>
-        <pie-chart data="pieChartData"
-                   options="COMPLETION_RATE_PIE_CHART_OPTIONS">
-        </pie-chart>
-        <p style="text-align: center;"><span ng-if="numPassersby" style="font-weight: bold;" class="protractor-test-num-passersby"><[numPassersby]></span><span ng-if="!numPassersby" style="font-weight: bold;">0</span> learners browsed this exploration. They left before reaching the second card.</p>
-
-        <h3>Card statistics</h3>
-        <div style="font-size: smaller" ng-if="lastUpdated">
-          <em>Last updated: <[getLocaleAbbreviatedDatetimeString(lastUpdated)]></em>
-        </div>
-        <div class="container">
-          <div class="row">
-            <div class="col-lg-12 col-md-12 col-sm-12">
-              <div state-graph-visualization graph-data="statsGraphData"
-                   node-fill="darkseagreen"
-                   highlight-states="highlightStates"
-                   state-stats="stateStats"
-                   on-click-function="onClickStateInStatsGraph">
-              </div>
-            </div>
-          </div>
-          <div class="container" ng-if="playthroughsAreAvailable">
-            <playthrough-issues></playthrough-issues>
-          </div>
-        </div>
-      </div>
-    </div>
-    <div ng-if="!explorationHasBeenVisited">
-      <em>
-        This exploration has not been visited by anyone yet, so there are no statistics to display.
-      </em>
-    </div>
-  </md-card>
-</div>
